Phase-II/backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.31
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
uvicorn==0.24.0
psycopg2-binary
pytest==7.4.3
httpx==0.25.2
alembic==1.13.1
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0